Backend/requirements.txt
Flask-SocketIO==5.5.1
python-socketio==5.12.1 
python-dotenv==1.0.1 
pymongo==4.11.1